+++ src/index.ts
@@ -67,8 +67,9 @@ export type {
   Task,
   TaskError,
   TaskPerformance,
-  WorkerStatistics
+  WorkerStatistics,
+  Writable
 } from './utility-types'
 export type { CircularArray } from './circular-array'
-export type { Queue } from './queue'
+export type { Deque } from './deque'
 export { availableParallelism } from './utils'